+++ CMakeLists.txt
@@ -2,10 +2,10 @@ cmake_minimum_required(VERSION 3.10)
 
 project(quadtree_pathfinding)
 
-add_library(quadtree_pathfinding SHARED quadtree_pathfinding.cpp)
-set_target_properties(quadtree_pathfinding
-    PROPERTIES PUBLIC_HEADER "quadtree_pathfinding.hpp")
 include_directories("3rd-party")
+file(GLOB QDPF_SOURCES qdpf.cpp internal/*.cpp)
+add_library(quadtree_pathfinding SHARED ${QDPF_SOURCES})
+set_target_properties(quadtree_pathfinding PROPERTIES PUBLIC_HEADER "qdpf.hpp")
 
 install(
     TARGETS quadtree_pathfinding
