+++ simple-example/CMakeLists.txt
@@ -8,7 +8,7 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS 1)
 include_directories("../3rd-party")
 include_directories("..")
 
-file(GLOB QDPF_SOURCES ../qdpf*.cpp)
+file(GLOB QDPF_SOURCES ../qdpf.cpp ../internal/*.cpp)
 add_library(qdpf ${QDPF_SOURCES})
 add_executable(quadtree-pathfinding-example main.cpp)
 target_link_libraries(quadtree-pathfinding-example qdpf)
